Allow reading web assets live from disk during development
