+++ bulletforcehax2_server/Cargo.toml
@@ -26,3 +26,6 @@ tower-http = { version = "0.4", features = ["cors"] }
 tracing = "0.1"
 tracing-appender = "0.2"
 tracing-subscriber = { version = "0.3", features = ["json"] }
+
+[features]
+read-from-disk = []
